Move interface, choose the better valuation set. Update tests.
